+++ libs/contract-plugin-converter/contracts/uberConverter.delay.test.ts
@@ -2,44 +2,21 @@ import hre from 'hardhat';
 import { jest } from '@jest/globals';
 import { createAndInitOracle, setOracleValueRepeat } from '@thecointech/contract-oracle/testHelpers.ts';
 import { ALL_PERMISSIONS, assignPlugin, buildAssignPluginRequest } from '@thecointech/contract-plugins';
-import { initAccounts, createAndInitTheCoin } from '@thecointech/contract-core/testHelpers.ts';
+import { createAndInitTheCoin, getSigners } from '@thecointech/contract-core/testHelpers.ts';
 import { buildUberTransfer } from '@thecointech/utilities/UberTransfer';
 import Decimal from 'decimal.js-light';
 import { DateTime, Duration } from 'luxon';
 import '@nomicfoundation/hardhat-ethers';
-import { EventLog, HDNodeWallet, Signer } from 'ethers';
-import { AccountName, getSigner } from '@thecointech/signers';
-import { SignerWithAddress } from '@nomicfoundation/hardhat-ethers/signers';
+import { EventLog } from 'ethers';
 
 const timeout = 10 * 60 * 1000;
 jest.setTimeout(timeout);
-
-
-const getSignerWithAddress = async (s: AccountName) => {
-  const signer = await getSigner(s);
-  return signer.connect(hre.ethers.provider) as unknown as SignerWithAddress;
-}
-const getSigners = async () => {
-  const Owner = await getSignerWithAddress("Owner");
-  const OracleUpdater = await getSignerWithAddress("OracleUpdater");
-  const Client1 = await getSignerWithAddress("Client1");
-  const Client2 = await getSignerWithAddress("Client2");
-  const BrokerCAD = await getSignerWithAddress("BrokerCAD");
-  return {
-    Owner,
-    OracleUpdater,
-    Client1,
-    Client2,
-    BrokerCAD
-  };
-}
 
 describe('Uberconverter delay tests', () => {
   it('Appropriately delays a transfer, and converts an appropriate amount at time', async () => {
 
     console.log("Start")
     const signers = await getSigners();
-    // const signers = initAccounts(await hre.ethers.getSigners());
     console.log("Signers initialized:", {
       owner: signers.Owner.address,
       client1: signers.Client1.address,
@@ -188,58 +165,58 @@ describe('Uberconverter delay tests', () => {
 });
 
 
-describe('Uberconverter current tests', () => {
+// describe('Uberconverter current tests', () => {
 
-  it('converts fiat to TheCoin for current transfers', async () => {
+//   it('converts fiat to TheCoin for current transfers', async () => {
 
-    const oriSigners = await hre.ethers.getSigners();
-    console.log(oriSigners[0].address);
-    const signers = await getSigners();
-    const UberConverter = await hre.ethers.getContractFactory('UberConverter');
-    const tcCore = await createAndInitTheCoin(signers.Owner);
-    const oracle = await createAndInitOracle(signers.OracleUpdater);
+//     const oriSigners = await hre.ethers.getSigners();
+//     console.log(oriSigners[0].address);
+//     const signers = await getSigners();
+//     const UberConverter = await hre.ethers.getContractFactory('UberConverter');
+//     const tcCore = await createAndInitTheCoin(signers.Owner);
+//     const oracle = await createAndInitOracle(signers.OracleUpdater);
 
-    // pass some $$$ to Client1
-    await tcCore.mintCoins(10000e6, signers.Owner, Date.now());
-    await tcCore.transfer(signers.Client1, 1000e6);
+//     // pass some $$$ to Client1
+//     await tcCore.mintCoins(10000e6, signers.Owner, Date.now());
+//     await tcCore.transfer(signers.Client1, 1000e6);
 
-    // Create plugin
-    const uber = await UberConverter.deploy();
-    await uber.initialize(tcCore, oracle);
+//     // Create plugin
+//     const uber = await UberConverter.deploy();
+//     await uber.initialize(tcCore, oracle);
 
-    // Assign to user, grant all permissions
-    const request = await buildAssignPluginRequest(signers.Client1, uber, ALL_PERMISSIONS);
-    await assignPlugin(tcCore, request);
+//     // Assign to user, grant all permissions
+//     const request = await buildAssignPluginRequest(signers.Client1, uber, ALL_PERMISSIONS);
+//     await assignPlugin(tcCore, request);
 
-    // Transfer $100 now.
-    const transfer = await buildUberTransfer(
-      signers.Client1,
-      signers.Client2,
-      new Decimal(100),
-      124,
-      DateTime.now(),
-    )
-    const initBalance = await tcCore.balanceOf(signers.Client1);
-    const r = await tcCore.uberTransfer(
-      transfer.chainId,
-      transfer.from,
-      transfer.to,
-      transfer.amount,
-      transfer.currency,
-      transfer.transferMillis,
-      transfer.signedMillis,
-      transfer.signature,
-    );
-    const receipt = await r.wait();
-    const Client1Balance = await tcCore.balanceOf(signers.Client1);
-    // If we transferred $100, that should have equalled 50C
-    expect(Number(initBalance - Client1Balance)).toEqual(50e6);
-    // Client2Balance should be the remainder
-    const Client2Balance = await tcCore.balanceOf(signers.Client2);
-    expect(Number(Client2Balance)).toEqual(50e6);
+//     // Transfer $100 now.
+//     const transfer = await buildUberTransfer(
+//       signers.Client1,
+//       signers.Client2,
+//       new Decimal(100),
+//       124,
+//       DateTime.now(),
+//     )
+//     const initBalance = await tcCore.balanceOf(signers.Client1);
+//     const r = await tcCore.uberTransfer(
+//       transfer.chainId,
+//       transfer.from,
+//       transfer.to,
+//       transfer.amount,
+//       transfer.currency,
+//       transfer.transferMillis,
+//       transfer.signedMillis,
+//       transfer.signature,
+//     );
+//     const receipt = await r.wait();
+//     const Client1Balance = await tcCore.balanceOf(signers.Client1);
+//     // If we transferred $100, that should have equalled 50C
+//     expect(Number(initBalance - Client1Balance)).toEqual(50e6);
+//     // Client2Balance should be the remainder
+//     const Client2Balance = await tcCore.balanceOf(signers.Client2);
+//     expect(Number(Client2Balance)).toEqual(50e6);
 
-    // The money was transferred, there should be logs!
-    expect(receipt.logs?.filter((e: EventLog) => e.eventName == "Transfer").length).toEqual(1);
-    expect(receipt.logs?.filter((e: EventLog) => e.eventName == "ExactTransfer").length).toEqual(1);
-  }, timeout);
-});
+//     // The money was transferred, there should be logs!
+//     expect(receipt.logs?.filter((e: EventLog) => e.eventName == "Transfer").length).toEqual(1);
+//     expect(receipt.logs?.filter((e: EventLog) => e.eventName == "ExactTransfer").length).toEqual(1);
+//   }, timeout);
+// });
